Organize workflow executors by action and move campaign email helpers into campaigns
