+++ apps/web/app/(ee)/api/e2e/trigger-workflow/[workflowId]/route.ts
@@ -1,6 +1,6 @@
 import { handleAndReturnErrorResponse } from "@/lib/api/errors";
-import { executeSendCampaignWorkflow } from "@/lib/api/workflows/execute-send-campaign-workflow";
 import { parseWorkflowConfig } from "@/lib/api/workflows/parse-workflow-config";
+import { executeSendCampaignWorkflow } from "@/lib/api/workflows/send-campaign/execute";
 import { withWorkspace } from "@/lib/auth";
 import { prisma } from "@/lib/prisma";
 import { WORKFLOW_ACTION_TYPES } from "@/lib/zod/schemas/workflows";
